#18 Added generator for procedure wrappers
Extended the interface generator to generate wrappers for procedures.
Modified the load generator to point to the wrappers. Regenerated the
parser.

TODO: modify the core to work with the new wrapper signature.
TODO: modify generated virtual methods & delegates to properly forward
to the new signature when they detect it’s a C function.