#21 Core interfaces are generated
This change enforces naming conventions of the C language binding on
the core by generating the interfaces.

It also generates the wrappers for the core procedures, which is
required because of #18.

Additionally, the change also moves interface/api into the core.

TODO: xml deserializer crashes on fast.xml
TODO: warning in generated wrappers
TODO: core is missing some procedures