#21 Fixes for various bugs and inconsistencies
A number of bugs and inconsistencies have been addressed in this commit:
 * Generated delegate routines did not provide a default value for a
returnvalue. When a callback was not implemented they would return an
uninitialized value which made it seem like initializers and
constructors were failing.
 * This has been fixed by having a delegate-callee check whether the
callback is implemented.
 * This caused another issue where delegates could not be resolved
during bootstrap. This has been solved by a much more efficient
approach than what was previously implemented. Instead of resolving the
delegate, a delegate-function will now simply point to the variable
provided by the language mapping that points to the delegate.
 * function::impl no longer points to the actual function but to the
wrapper. A number of locations in the code still expected impl to
contain the function which caused various crashes.

A new bug has been identified for which an issue must be created. The
issue has to do with resolving constants in expressions. A new testcase
tc_bitmask01.hyve captures the bug. The bug also surfaces in the
os.hyve file.

This commit fixed #21