#16 Added creation of mock HTML files
Note that these html files do already have a link to the appropriate js file in the same folder.

The next step is to display the information of these serialized objects in an appropriate manner.
We have to include a magic js script that will parse this Json and display it in a table.