#164 OS X support (1st version)
This commit contains a port to OS X along with several bug fixes (of
which most surfaced during porting). A list of changes:

 * Use clang on OS X
 * Change cx_*mutexNew functions. It seems that on OS X the address
passed to pthread_*lock_init is stored somewhere.
 * Add custom implementation of strdup
 * Add code that detects darwin
 * Linking with librt is not required on OS X
 * Anonymous observable objects were not properly initialized
 * strcasecmp replaced with custom implementation
 * Implemented alternative to clock_gettime on OS X
 * c_common is now a library, instead of a generator. Linking with a
library that doesn’t confirm lib*.so caused issues.
 * The flex library on OS X is called ‘l’ instead of ‘fl’
 * cxsh would crash on displaying objects with an anonymous type

The tc_function03 testcase is crashing on OS X (the rest of the test
cases pass).