#14 Implemented iterator functionality for arrays and sequences
A minor issue still exists with lists, which has to do with how values
smaller than a word size are stored in a list.

The ‘current’ member of the iterator expects an address to the value,
however for the above mentioned values the list doesn’t store an
address to the value, but the value itself.

Further TODO’s is implement initializers and build testcases.