#184 #192 In json.c, forwarded serialize-base to serialize-value; extracted methods for primitive serialization
Before, serialize-base was calling serialize members instead of serialize-value. This also helps make the code slightly more concise by letting serialize-composite handle the addition of curly braces.

Because serialization of primitives was handled in the same method all together, the code was not very readable and was difficult to maintain. Now, serialize-primitive will forward to one of several new static methods for each primitive kind.