#211 removed support for 0 parentState
The type::parentState member indicates the state a parent must be
before an object can be declared and defined. When the parentState was
not explicitly set, its value was zero, which is invalid. The
parentState is a bitmask that must be either DECLARED, DEFINED or
DECLARED|DEFINED. Added code to the initializer of a type that sets the
default parentState to DECLARED|DEFINED.