#393 Don't run expression until return is pressed
The old code called corto_expr to validate whether an expression was
valid, and apply colors/resolve members accordingly.

corto_expr runs the expression however, which means that if an
expression has side effects, it could be executed many times before a
user presses enter.

The code now uses a (new) function of the language parser, that just
resolves the type of an expression without running it.