#361 finished orphaned objects, make ownership target aware
- New ‘attr’ field for types that can mandate with which attributes an
object of a specific type is created. Using this field, it can be
guaranteed that a target instance is always created as a scoped object.
- Support for observable members, which instantiates the member as an
orphaned object with the parent set to the containing object, so that
updates on the member are forwarded correctly
- Orphaned objects now do not have the CORTO_DECLARED state. This is
used by corto to ensure that they do not emit DEFINE and DELETE events.
- Ownership of orphaned objects is determined by the owner of the
parent object. ownerof will look for the first parent that has the
CORTO_DECLARED state. If no such parent is found, it returns NULL.
- corto_update* does not throw an error for an instance of a target
type, even if that object is not owned by the current thread. It is the
responsibility of the application to only set the appropriate target or
actual field. This is encapsulated when using the generated C-binding
functions.