#341 rewrite corto_resolve so that it won't resolve parents separately
This behavior caused issues with the security framework, because while
a user may have access to a child, he/she may not have access to its
parent. Because corto_resolve was resolving the parent first, and the
child afterwards, the resolve would fail even though the user had
access.

As a nice side-effect, this also results in fewer resumes when no
object is resolved, whereas previously parent objects would be
individually resumed (and suspended) during the resolve.