#341 implement new loader resume function
Because the new corto_resume (and corto_resolve) functions do no longer resolve each element in an identifier separately, objects that were located in packages could not be resolved because their parent (the package) was not loaded as part of resolving the object.

To address this, a new resume function has been introduced in the package loader that will cascade up an identifier until it has found a package, and consecutively attempts to resolve the object in the resolved package.