#685 fix observercrash on shutdown
The observable administration was called when corto_delete was invoked
on an object. However, if the object contained child objects and the
observer was observing with ON_SCOPE or ON_TREE, those child objects
would attempt to notify the observer through their parent. To enable
this, the observer administration needs to be deleted only at the
moment the object is actually freed from the heap.