#341 rename result to record
The corto/vstore/result identifier was used for the type that describes
objects that do not live in the in-memory store. “result” is too
generic though, and it doesn’t describe well what its function is (when
used in a subscriber, nothing is “returned”). Therefore the typename is
changed to “record”, which is more appropriate.